+++ pollux/lib/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [pollux-v0.42.0](https://github.com/input-output-hk/atala-prism-building-blocks/compare/pollux-v0.41.0...pollux-v0.42.0) (2023-03-16)
+
+
+### Features
+
+* **pollux:** update credential schema logic ([#450](https://github.com/input-output-hk/atala-prism-building-blocks/issues/450)) ATL-3164 ([6e22bfc](https://github.com/input-output-hk/atala-prism-building-blocks/commit/6e22bfcd582b4e33611a91d9ce1339a222778d25)), closes [#2](https://github.com/input-output-hk/atala-prism-building-blocks/issues/2)
+
+# [pollux-v0.41.0](https://github.com/input-output-hk/atala-prism-building-blocks/compare/pollux-v0.40.0...pollux-v0.41.0) (2023-03-15)
+
+
+### Features
+
+* **pollux:** Add ignoreWithZeroRetries filter to SQL queries ([#445](https://github.com/input-output-hk/atala-prism-building-blocks/issues/445)) ([d477c68](https://github.com/input-output-hk/atala-prism-building-blocks/commit/d477c683944a30125a1caa6b5f66caa156e7f70a))
+
 # [pollux-v0.40.0](https://github.com/input-output-hk/atala-prism-building-blocks/compare/pollux-v0.39.0...pollux-v0.40.0) (2023-03-10)